# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Retriva Gateway API
  description: |
    Retriva Gateway acts as the Backend-for-Frontend (BFF) for the Retriva WebUI
    and provides a public-facing API for external integrations.

    It exposes:
    * `/gateway/*`: Backend-for-Frontend (BFF) routes specifically designed for the WebUI.
    * `/api/v2/*`: Public-facing modern API routes.
  version: 1.0.0
  contact:
    name: Andrea Marson
    email: am.dev.75@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8002
    description: Retriva Gateway API
...
paths:
...
  /gateway/internal/sources/{source_id}/runs/{run_id}/heartbeat:
    post:
      tags: [Internal]
      summary: Worker heartbeat
      operationId: workerHeartbeat
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/ServiceToken"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: OK
  /gateway/internal/sources/{source_id}/runs/{run_id}/events:
    post:
      tags: [Internal]
      summary: Batch item state updates
      operationId: runEvents
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/ServiceToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventsRequest"
      responses:
        "200":
          description: Events processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: integer }
  /gateway/internal/sources/{source_id}/runs/{run_id}/complete:
    post:
      tags: [Internal]
      summary: Complete or fail a run
      operationId: completeRun
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/ServiceToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRequest"
      responses:
        "200":
          description: Run completed
  /gateway/internal/sources/{source_id}/ingestion-session:
    post:
      tags: [Internal]
      summary: Issue scoped ingestion context
      operationId: createIngestionSession
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/ServiceToken"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestionSessionRequest"
      responses:
        "200":
          description: Ingestion session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionSessionResponse"
  /gateway/internal/sources/{source_id}/checkpoint:
    get:
      tags: [Internal]
      summary: Get latest source checkpoint
      operationId: getSourceCheckpoint
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/ServiceToken"
      responses:
        "200":
          description: Latest source checkpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceCheckpointResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Source or checkpoint not found
components:
  schemas:
...
    ItemStatus:
      type: string
      enum: [pending, indexed, deleted, error]
...
    HeartbeatRequest:
      type: object
      properties:
        processed_items: { type: integer, nullable: true }
        failed_items: { type: integer, nullable: true }
        skipped_items: { type: integer, nullable: true }
    ItemEvent:
      type: object
      required: [source_item_id]
      properties:
        source_item_id: { type: string }
        source_revision: { type: string, nullable: true }
        retriva_doc_id: { type: string, nullable: true }
        content_hash: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/ItemStatus"
    EventsRequest:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ItemEvent"
    CompleteRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [completed, failed] }
        error_code: { type: string, nullable: true }
        error_message: { type: string, nullable: true }
        checkpoint: { type: object, nullable: true, additionalProperties: {} }
        processed_items: { type: integer, nullable: true }
        failed_items: { type: integer, nullable: true }
        skipped_items: { type: integer, nullable: true }
    IngestionSessionRequest:
      type: object
      properties:
        run_id: { type: string, nullable: true }
    IngestionSessionResponse:
      type: object
      required: [kb_id, tenant_id, source_id, batch_metadata]
      properties:
        kb_id: { type: string }
        tenant_id: { type: string }
        source_id: { type: string }
        run_id: { type: string, nullable: true }
        batch_metadata: { type: object, additionalProperties: {} }
    SourceCheckpointResponse:
      type: object
      required: [source_id, checkpoint_type, updated_at]
      properties:
        source_id:
          type: string
        checkpoint_type:
          type: string
          description: Connector-specific checkpoint type, e.g. mediawiki_recentchanges
        baseline_started_at:
          type: string
          format: date-time
          nullable: true
        last_seen_timestamp:
          type: string
          format: date-time
          nullable: true
        last_seen_rcid:
          type: integer
          nullable: true
        cursor:
          type: object
          nullable: true
          additionalProperties: {}
        updated_at:
          type: string
          format: date-time
  parameters:
    SourceId:
      name: source_id
      in: path
      required: true
      schema: { type: string }
      description: Connected source ID (e.g. src_a1b2c3d4e5f6)
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
      description: Sync run ID (e.g. run_a1b2c3d4e5f6)
    ServiceToken:
      name: X-Service-Token
      in: header
      required: false
      schema: { type: string }
      description: Service authentication token for internal endpoints